+++ src/modules/settings/components/updateDaoContracts/index.ts
@@ -1,1 +1,1 @@
-export { UpdateDaoContracts, type IUpdateDaoContractsProps } from './updateDaoContarcts';
+export { UpdateDaoContracts, type IUpdateDaoContractsProps } from './updateDaoContracts';
